Add support for buildargs.
The docker api requires the buildargs parameter to be a valid (escaped)
json hash. So, if the `buildargs` key (in create section of yaml file)
is a hash, convert it to json before invoking `build_from_dir`.

if`buildargs` is a string it will be passed through as-is, so it should
be a valid json hash (e.g., `'{"foo":"bar"}'`).